+++ asd/rendiconti/ch.banana.it.extension.asd.rendicontogestionale.js
@@ -1,4 +1,4 @@
-// Copyright [2024] [Banana.ch SA - Lugano Switzerland]
+// Copyright [2025] [Banana.ch SA - Lugano Switzerland]
 // 
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -14,7 +14,7 @@
 //
 // @id = ch.banana.it.extension.asd.rendicontogestionale
 // @api = 1.0
-// @pubdate = 2024-01-05
+// @pubdate = 2025-05-13
 // @publisher = Banana.ch SA
 // @description = 2. Rendiconto gestionale ASD
 // @task = app.command
@@ -97,7 +97,7 @@ function stampaReportNormale(banDoc, paramReport, stylesheet) {
    bReport.validateGroups_IncomeExpenses(paramReport.userParam.column, paramReport.reportStructure);
    bReport.loadBalances();
    bReport.calculateTotals(["currentAmount", "previousAmount"]);
-   bReport.formatValues(["currentAmount", "previousAmount"]);
+   bReport.formatValues(["currentAmount", "previousAmount"], paramReport.userParam.excludedecimals);
 
    var report = printRendicontoModB(banDoc, paramReport.userParam, bReport, stylesheet);
 
@@ -1493,6 +1493,18 @@ function convertParam(userParam) {
    convertedParam.data.push(currentParam);
 
    currentParam = {};
+   currentParam.name = 'excludedecimals';
+   currentParam.parentObject = 'report_group';
+   currentParam.title = 'Arrotonda gli importi all’intero (senza cifre decimali)';
+   currentParam.type = 'bool';
+   currentParam.value = userParam.excludedecimals ? true : false;
+   currentParam.defaultvalue = false;
+   currentParam.readValue = function() {
+    userParam.excludedecimals = this.value;
+   }
+   convertedParam.data.push(currentParam);
+
+   currentParam = {};
    currentParam.name = 'styles';
    currentParam.title = 'Stili';
    currentParam.type = 'string';
@@ -1552,6 +1564,7 @@ function initUserParam() {
    userParam.textbegin = '';
    userParam.column = 'Gr1'; //per il raggruppamento si usa sempre la colonna Gr1
    userParam.printcolumn = true;
+   userParam.excludedecimals = false;
    userParam.finalnotes = '';
    userParam.colorheadertable = '#337ab7';
    userParam.stampareportcontrollo = false;